Added license and removed dnnl_graph_compiler.hpp
